// src/types/indexeddb.types.ts
export interface IndexedDBColumn {
  name: string
  keyPath: string
  options?: IDBIndexParameters
}

export interface IndexedDBStore {
  id: IDBObjectStoreParameters
  name: string
  indices?: IndexedDBColumn[]
}

export interface IndexedDBConfig {
  databaseName: string
  version: number
  stores: IndexedDBStore[]
}

export interface TransactionOptions {
  storeName: string
  dbMode: IDBTransactionMode
  error: (e: Event) => any
  complete: (e: Event) => any
  abort?: any
}